Proposal for test setup simplification
