+++ packages/deployer/utils/tests.js
@@ -1,0 +1,37 @@
+const { deploySystem, getProxyAddress, getRouterAddress } = require('@synthetixio/deployer/utils/deployments');
+const { takeSnapshot, restoreSnapshot } = require('@synthetixio/core-js/utils/rpc');
+
+function bootstrap(initializer) {
+  let snapshotId;
+
+  const deploymentInfo = {
+    network: hre.config.defaultNetwork,
+    instance: 'test',
+  };
+
+  const proxyAddress = getProxyAddress(deploymentInfo);
+  const routerAddress = getRouterAddress(deploymentInfo);
+
+  before('take a snapshot', async () => {
+    snapshotId = await takeSnapshot(hre.ethers.provider);
+  });
+
+  before('deploy system', async () => {
+    await deploySystem(deploymentInfo, { clear: true });
+  });
+
+  before('initialize system', async () => {
+    await initializer(deploymentInfo);
+  });
+
+  after('restore the snapshot', async () => {
+    await restoreSnapshot(snapshotId, hre.ethers.provider);
+  });
+
+  return { deploymentInfo, proxyAddress, routerAddress };
+};
+
+module.exports = {
+  bootstrap,
+  deploySystem,
+};
